refactor: re-organize project structure, add new paths to gitignore
