+++ .gitignore
@@ -1,3 +1,7 @@
+target/
+!**/src/main/**/target/
+!**/src/test/**/target/
+
 ## IntelliJ
 out/
 .idea/
