+++ src/main/java/utils/DateFormatter.java
@@ -1,4 +1,4 @@
-package src.utils;
+package utils;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
